requirements-database.txt
# =====================================
# مكتبات قاعدة البيانات
# =====================================

# SQLite (مدمج في Python - لا يحتاج تثبيت)

# MySQL
pymysql>=1.1.0

# PostgreSQL
psycopg2-binary>=2.9.9

# SQLAlchemy (مطلوب للجميع)
Flask-SQLAlchemy>=3.1.1
SQLAlchemy>=2.0.23

# أدوات إضافية
python-dotenv>=1.0.0  # لتحميل ملف .env
